Refactor to use correct types instead of String everywhere

+++ src/gorske.rs
@@ -10,21 +10,21 @@ pub fn get_gorske() -> f32 {
 }
 
 /// Calculates our wasted cost in Dollarydoos
-pub fn calculate_cost(hours: f32, hourly_rate: f32, staff_count: i32) -> f32 {
+pub fn calculate_cost(hours: f32, hourly_rate: f32, staff_count: f32) -> f32 {
     hours * hourly_rate * staff_count as f32 * 52.0
 }
 
 /// Converts a yearly salary into a hourly salary
-pub fn calculate_hourly_salary(yearly_salary: i32) -> f32 {
+pub fn calculate_hourly_salary(yearly_salary: f32) -> f32 {
     yearly_salary as f32 / 52.0 / 40.0
 }
 
 /// Calculates a given cost and outputs in Gorske
-pub fn calculate_gorske_cost(hours: f32, hourly_rate: f32, staff_count: i32) -> f32 {
+pub fn calculate_gorske_cost(hours: f32, hourly_rate: f32, staff_count: f32) -> f32 {
     calculate_cost(hours, hourly_rate, staff_count) / get_gorske()
 }
 
 /// Calculates a given cost and outputs in BigMac's
-pub fn calculate_bigmac_cost(hours: f32, hourly_rate: f32, staff_count: i32) -> f32 {
+pub fn calculate_bigmac_cost(hours: f32, hourly_rate: f32, staff_count: f32) -> f32 {
     calculate_cost(hours, hourly_rate, staff_count) / BIGMAC_PRICE
 }
